+++ src/Experimental/RAGTools.jl
@@ -1,0 +1,16 @@
+module RAGTools
+
+@warn """
+RAGTools functionality has moved to a dedicated package!
+
+The PromptingTools.Experimental.RAGTools module has been moved to the separate RAGTools.jl package.
+
+To migrate your code:
+1. Add RAGTools.jl: using Pkg; Pkg.add("RAGTools")
+2. Update imports: using RAGTools (instead of using PromptingTools.Experimental.RAGTools)
+3. All function names and APIs remain the same
+
+For more information, visit: https://github.com/JuliaGenAI/RAGTools.jl
+""" maxlog=1
+
+end # module RAGTools